backend/requirement.txt
Django>=4.2
djangorestframework>=3.15
django-cors-headers>=4.1
psycopg2-binary>=2.9 
